+++ gulpfile.js
@@ -30,18 +30,8 @@ gulp.task('pack_demo', function(cb) {
     });
 });
 
-gulp.task('less_component', function(cb) {
-    gulp.src(['./src/**/*.less'])
-        .pipe(sourcemaps.init())
-        .pipe(less())
-        .pipe(sourcemaps.write('.'))
-        .pipe(gulp.dest('./src'));
-    console.info('###### less_component done ######');
-    cb();
-});
-
 gulp.task('less_demo', function(cb) {
-    gulp.src(['./demo/**/*.less'])
+    gulp.src(['./src/*.less','./demo/**/*.less'])
         .pipe(sourcemaps.init())
         .pipe(less())
         .pipe(concat('demo.css'))
@@ -69,20 +59,18 @@ gulp.task('reload_by_demo_css', ['less_demo'], function () {
 
 gulp.task('server', [
     'pack_demo',
-    'less_component',
     'less_demo'
 ], function() {
     browserSync({
         server: {
             baseDir: './'
         },
-
         open: 'external'
     });
 
     gulp.watch(['src/**/*.js', 'demo/**/*.js'], ['reload_by_js']);
 
-    gulp.watch('src/**/*.less', ['reload_by_component_css']);
+    gulp.watch('src/**/*.less', ['reload_by_demo_css']);
 
     gulp.watch('demo/**/*.less', ['reload_by_demo_css']);